# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/members/{id}/images": {
      "get": {
        "operationId": "ImageController_getMemberImages_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all payment screenshots and profile images for a member",
        "tags": ["Admin - Images"]
      },
      "post": {
        "description": "Upload an image file or provide an imageUrl. File upload is preferred.",
        "operationId": "ImageController_addImage_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID (UUID)",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image file (JPEG, PNG, GIF, WEBP)"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["PROFILE", "PAYMENT_SCREENSHOT"],
                    "description": "Image type"
                  },
                  "description": {
                    "type": "string",
                    "description": "Image description (optional)"
                  },
                  "imageUrl": {
                    "type": "string",
                    "description": "Image URL (optional if file is uploaded)"
                  }
                },
                "required": ["type"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add new image (profile or payment screenshot)",
        "tags": ["Admin - Images"]
      }
    },
    "/api/v1/images/{id}": {
      "patch": {
        "operationId": "ImageController_updateImage_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Image ID (UUID or profile-{memberId})",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateImageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update image metadata (description)",
        "tags": ["Admin - Images"]
      },
      "delete": {
        "operationId": "ImageController_deleteImage_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Image ID (UUID or profile-{memberId})",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete single image",
        "tags": ["Admin - Images"]
      }
    },
...
  },
  "info": {
    "title": "Gym Mini Check-in System API",
    "description": "API documentation for Gym Mini Check-in System",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateImageDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Image description or note",
            "example": "Updated payment screenshot"
          }
        }
      },
...
    }
  }
}